UPDATE: Adds GetCart by usedrId endpoint added

+++ Ecom.Services.ShoppingCartApi/Models/Dtos/CouponDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Code { get; set; }
-        public double DiscountAmount { get; set; }
+        public decimal DiscountAmount { get; set; }
         public int MinAmount { get; set; }
     }
 }
